requirements.txt
APScheduler==3.6.3
async-timeout==4.0.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.9.24
Deprecated==1.2.13
importlib-metadata==5.0.0
packaging==21.3
pyparsing==3.0.9
python-telegram-bot==13.14
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
six==1.16.0
tornado==6.1
typing-extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
wrapt==1.14.1
zipp==3.9.0
